Reorganize python and c++ examples following dynet
