+++ examples/cpp/dense/FactorSequence.h
