+++ examples/cpp/dense/simple_grid.cpp
