+++ examples/cpp/logic/Makefile
@@ -1,8 +1,8 @@
 OBJS =
 CC = g++
 DEBUG = -g
-INCLUDES = -I../.. 
-LIBS = -L/usr/local/lib/ -L../../ad3/
+INCLUDES = -I../../../ 
+LIBS = -L/usr/local/lib/ -L../../../ad3/
 CFLAGS = -O3 -Wall -Wno-sign-compare -c -fmessage-length=0 $(INCLUDES)
 LFLAGS = $(LIBS) -lad3
