+++ examples/cpp/logic/simple_coref.cpp
