+++ examples/cpp/parsing/FactorHeadAutomaton.h
