+++ examples/cpp/parsing/FactorTree.cpp
