+++ examples/cpp/parsing/FactorTree.h
