+++ examples/cpp/parsing/simple_parser.cpp
