+++ examples/cpp/summarization/FactorCompressionBudget.h
