+++ examples/cpp/summarization/FactorGeneralTree.h
