+++ examples/cpp/summarization/FactorGeneralTreeCounts.h
