+++ examples/cpp/summarization/FactorSequenceBudget.h
