+++ examples/python/example_binary_tree.py
