+++ examples/python/example_compression_budget.py
