+++ examples/python/example_grid.py
