+++ examples/python/example_grid_diversity.py
