+++ examples/python/example_knapsack.py
