+++ examples/python/example_tree_budget.py
